CHANGES
So, I made 2 folders, New And Old, NEw is the current version im working on and today i made that borrow and return thing work,like somewhat.  Old is the previous persion of the project and I dont know if we wanna keep that or not so i just kept it obviously. Itll be like backup only lol